LPS-56703 Move BlogsPortletInstanceSettings
